+++ src/basic/__tests__/basic.test.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import { describe, expect, test } from 'vitest';
 import { act, fireEvent, render, renderHook, screen, within } from '@testing-library/react';
-import { CartPage } from '../../basic/components/CartPage';
-import { AdminPage } from "../../basic/components/AdminPage";
+import { CartPage } from '../../refactoring/components/CartPage';
+import { AdminPage } from "../../refactoring/components/AdminPage";
 import { CartItem, Coupon, Product } from '../../types';
-import { useCart, useCoupons, useProducts } from "../../basic/hooks";
-import * as cartUtils from "../../basic/hooks/utils/cartUtils";
+import { useCart, useCoupons, useProducts } from "../../refactoring/hooks";
+import * as cartUtils from "../../refactoring/hooks/utils/cartUtils";
 
 const mockProducts: Product[] = [
   {
