+++ src/refactoring/hooks/useCart.ts
@@ -7,19 +7,42 @@ export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
 
-  const addToCart = (product: Product) => {};
+  const getRemainingStock = (product: Product) => {
+    const cartItem = cart.find(item => item.product.id === product.id);
+    return product.stock - (cartItem?.quantity || 0);
+  };
 
-  const removeFromCart = (productId: string) => {};
+  const addToCart = (product: Product) => {
+    const remainingStock = getRemainingStock(product);
+    if (remainingStock <= 0) return;
 
-  const updateQuantity = (productId: string, newQuantity: number) => {};
+    setCart(prevCart => {
+      const existingItem = prevCart.find(item => item.product.id === product.id);
+      if (existingItem) {
+        return prevCart.map(item =>
+          item.product.id === product.id
+            ? { ...item, quantity: Math.min(item.quantity + 1, product.stock) }
+            : item
+        );
+      }
+      return [...prevCart, { product, quantity: 1 }];
+    });
+  };
 
-  const applyCoupon = (coupon: Coupon) => {};
+  const removeFromCart = (productId: string) => {
+    setCart(prevCart => prevCart.filter(item => item.product.id !== productId));
+  };
 
-  const calculateTotal = () => ({
-    totalBeforeDiscount: 0,
-    totalAfterDiscount: 0,
-    totalDiscount: 0,
-  })
+  const updateQuantity = (productId: string, newQuantity: number) => {
+    setCart(prevCart => updateCartItemQuantity(prevCart,productId,newQuantity)
+  );
+  };
+  const applyCoupon = (coupon: Coupon) => {
+    setSelectedCoupon(coupon);
+  };
+  const calculateTotal = () => {
+    return calculateCartTotal(cart,selectedCoupon)
+  };
 
   return {
     cart,
