+++ src/refactoring/hooks/useCoupon.ts
@@ -2,5 +2,10 @@ import { Coupon } from "../../types.ts";
 import { useState } from "react";
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
-  return { coupons: [], addCoupon: () => undefined };
+  const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
+
+  function handleCouponAdd(newCoupon: Coupon) {
+    setCoupons(prevCoupons => [...prevCoupons, newCoupon]);
+  };
+  return { coupons, addCoupon: handleCouponAdd };
 };
