+++ src/refactoring/hooks/utils/cartUtils.ts
@@ -1,21 +1,63 @@
 import { CartItem, Coupon } from "../../../types";
 
 export const calculateItemTotal = (item: CartItem) => {
-  return 0;
+  const { quantity } = item;
+  const discount = item.product.discounts.reduce((maxDiscount, d) => {
+    return quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount;
+  }, 0);
+
+  return item.product.price * item.quantity * (1 - discount);
 };
 
 export const getMaxApplicableDiscount = (item: CartItem) => {
-  return 0;
+  const { discounts } = item.product;
+  const { quantity } = item;
+  let appliedDiscount = 0;
+  for (const discount of discounts) {
+    if (quantity >= discount.quantity) {
+      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+    }
+  }
+  return appliedDiscount;
 };
 
 export const calculateCartTotal = (cart: CartItem[], selectedCoupon: Coupon | null) => {
+  const totalBeforeDiscount = cart.reduce((total, item) => {
+    return total + item.product.price * item.quantity;
+  }, 0);
+
+  const totalAfterDiscount = cart.reduce((total, item) => {
+    return total + calculateItemTotal(item);
+  }, 0);
+
+  let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+
+  // 쿠폰 적용
+  let finalTotalAfterDiscount = totalAfterDiscount;
+  if (selectedCoupon) {
+    if (selectedCoupon.discountType === 'amount') {
+      finalTotalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue);
+    } else {
+      finalTotalAfterDiscount *= (1 - selectedCoupon.discountValue / 100);
+    }
+    totalDiscount = totalBeforeDiscount - finalTotalAfterDiscount;
+  }
+
   return {
-    totalBeforeDiscount: 0,
-    totalAfterDiscount: 0,
-    totalDiscount: 0,
+    totalBeforeDiscount: Math.round(totalBeforeDiscount),
+    totalAfterDiscount: Math.round(finalTotalAfterDiscount),
+    totalDiscount: Math.round(totalDiscount),
   };
 };
 
 export const updateCartItemQuantity = (cart: CartItem[], productId: string, newQuantity: number): CartItem[] => {
-  return []
+  return cart.map(item => {
+      if (item.product.id === productId) {
+        const maxQuantity = item.product.stock;
+        const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
+        return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
+      }
+      return item;
+    }).filter((item): item is CartItem => item !== null)
+  
 };
